Converted all commands to classes
Converted all the commands to classes which implement their specific types.

Also, added a new property of CommandType which can be "regular" or "infraction".

Also implemented logging where client.emit is used to emit an loggerCreate event which passes through an embed using which iit can log changes or events.